# Conflicts:
#	Robot'em Up/Assets/Scenes/VSD/VSD3_TUTO1.unity
#	Robot'em Up/ProjectSettings/ProBuilderSettings.json

+++ Robot'em Up/Assets/Scripts/AI Behaviours/Boss Enemy/EnemyBoss.cs
@@ -652,7 +652,7 @@ public class EnemyBoss : PawnController, IHitable
     }
 
 
-    public void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, int _damages, DamageSource _source, Vector3 _bumpModificators = default)
+    public void OnHit(BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, float _damages, DamageSource _source, Vector3 _bumpModificators = default)
     {
         if (_ball != null)
         {
